Fix compenent commit
