+++ components/Sunshine/Sunshine.tsx
@@ -5,6 +5,227 @@ import { IUIComponent, Range } from '../../utils/types'
 import { Box } from '../Box'
 import React from 'react'
 
+/*
+                                                                                                                                                                                                                                                                ... ..
+                                                                                                                                                                                                                                                                ...   ..
+                                                                                                                                                                                                                                                               ..      .
+                                                                                                                                                                                                                                                               .       .
+                                                                         ...                                                                                                                                                                                    ..    ..
+                    .                   .                              .. . ..                                                                                                                                                                                   ... ..                                                                                                                                                                                                                   ....
+                ...  ..                 .                             ..     .                                                                                                                                                                                                                                                                                                                                                                                                          .. . .
+                        .               .                             .       .                                                                                                                                                                                                                                                                                                                                                                                                         .     .
+                         .             .                              ..      .                                                                                                                                                               ...                                                                                                                                                                                                                                              .
+                         .            .        ..... .                ..      .                                                                                                                                                               .  ....                                                                                                                                                                                                                                   .     .
+                       . .            .        .       .... ..          .   ..                                                                                                                                                          .  .. ..     ..                                                                                                                                                                                           .                           .          . ...
+                      .               .        .                         ....                                                                                                                                                       .. .              .                                                                                                                                                                                          .                  .   . . ...
+                     .               .         .                                                                                                                                                                                    .                  .                                                                                                                                                                                         .                  ...
+                 .. . ..              .        .                                                                                                                                                                                   .                   .                                                                                                                                                                                        .                   .
+                 .      ..          . .        ...                                                                                                                                                                                 .                   .                                                                                                                                                                                        .                   .
+                         .         . .            ..                                                                                                                                                                               .                   .                                                                                                                                                                                        .                   .
+                         .       ..  .              ..                                                                                                                                                                             .                  .                                                                                                                                                                                        .                   .
+                         .       .   .                .                                                                                                                                                                            .                  .                                                                                                                                                                                        .                   .
+            ..          .            .        .       ..                                                                                                                                                                           .                  .                                                                                                                                                                                        .         ..        .
+            ..        ..             ..        ..      .                                                                                                                                                                            .                .                                                                                                                                                                                          .. . .. . .        .
+              . ..  ..                           . .....                                                                                                                                                                            .               ..                                                                                                                                                                                                   .         .....
+                                                                                                                                                                                                                                    ..             .                                                                                                                                                                                                     .              ..
+                                                                                                                                                                                                                                      ..         ..                                                                                                                                                                                                     .                 .
+                                                                                                                                                                                                                                        .. . ....                                                                                                                                                                                                       .         .       .
+                                                                                                                                                                                                                                            ..                                                                                                                                                                                                          .         .      ..
+                                                                                                                                                                                                                                                                                                                                                                                                                                                        .         ..    ..
+                                                                                                                                                                                                                                                                                                                                                                                                                                                                   ... ..
+
+
+
+                                                                             ┌────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┬───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┐
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                            1                                                                                                       │                 000000000                                                                                                                                                     │
+                                                                             │                                                           11                8                                              1                                       │               00        0000                                                              1                                                                                   │
+                                                                             │             1    77                                       11             888 88888                                        11                99                     │             00             000                                                            1                                                   22222                           │
+                                                                             │            11      77777777                              1 1            888      88                                      111             999 999                   │            0                  00                                                         11                                                   2    2                          │
+                                                                             │            1              7                              1 1              8       8                                     1  1            9      9 9                 │           0                    00                                                      1111                                                   2     2                         │
+                                                                             │            1             7                              1  1              88     88                                    11  1           9        99                 │          0                      00                                                    11 1                                                         22                         │
+                                                                             │           11             7                              1  1               88  88                                     11   1            9     999                  │         0                        0                                                  11   1                                                         2                          │
+                                                                             │           11             7                            11   1                888                                      1     1            9999999 9                  │         0                         0                                               11     1                                                         2                          │
+                                                                             │          1 1             7                           11    1              8888                                      11     1                    9                  │         0                         0                                              1       1                                                        2                           │
+                                                                             │         111              77777777                    1    1             88   8                                    111      1                   9                   │        0                          0                                            1         1                                                       22                           │
+                                                                             │         1 1        7777777                                1            88    88                                            1                   9                   │        0                          0                                         111          1                                                      22                            │
+                                                                             │           1              7                                1            88     8                                            1                   9                   │        0                          0                                      111             1                                                    222                             │
+                                                                             │           1              7                                1             8    8                                             1                  99                   │         0                         0                                                     1                                                   222                               │
+                                                                             │           1             7                                 1              88888                                             1               9999                    │         00                        0                                                     1                                                  22               2                 │
+                                                                             │           1             7                                                                                                                                          │          0                       00                                                     1                                                 22222222          22                │
+                                                                             │           1             7                                                                                                                                          │           0                      0                                               1111   1          1                                    2222     222      222                 │
+                                                                             │          1                                                                                                                                                         │            0                   00                                           11111111 1111111 1 1 1111                                   2           2222222                   │
+                                                                             │                                                                                                                                                                    │             0                 00                                                                                                                                              │
+                                                                             │                                                                                                                                                                    │              000             0                                                                                                                                                │
+                                                                             │                                                                                                                                                                    │                 0000000000000                                                                                                                                                 │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                   33                          │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 33   33                       │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                3       33                     │
+                                                                             │            1          666                                                                                                                                          │                                                                                                                                               3         3                     │
+                                                                             │            1         6                                                                                                                                             │                                                                                                                                                         3                     │
+                                                                             │           11        6                                                                                                                                              │                                                                                                                                                        3                      │
+                                                                             │          111       6                                                                                                                                               │                                                                                                                                                        3                      │
+                                                                             │          1 1      6                                                                                                                                                │                                                                                                                                                       3                       │
+                                                                             │         1 1      66                                                                                                                                                │                                                                                                                                                    333                        │
+                                                                             │        1  1      6                                                                                                                                                 │                                                                                                                                                  3333                         │
+                                                                             │       1   1      6                                                                                                                                                 │                                                                                                                                                3333 3333                      │
+                                                                             │      1    1     6  66666                                                                                                                                           │                                                                                                                                                        3                      │
+                                                                             │     1    1      6 6    666                                                                                                                                         │                                                                                                                                                        3                      │
+                                                                             │          1      66       6                                                                                                                                         │                                                                                                                                                        3                      │
+                                                                             │          1      66       6                                                                                                                                         │                                                                                                                                                       3                       │
+                                                                             │          1      6        6                                                                                                                                         │                                                                                                                                                      33                       │
+                                                                             │          1       6       6                                                                                                                                         │                                                                                                                                                      3                        │
+                                                                             │          1       666  666                                                                                                                                          │                                                                                                                                               33    33                        │
+                                                                             │          1         666                                                                                                                                             │                                                                                                                                                33333                          │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                  4                            │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 44                            │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 4                             │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                4                              │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                4                              │
+                                                                             │                                                                                                                                                                    │                                                                                                                                               4                               │
+                                                                             │            1                                                                                                                                                       │                                                                                                                                              44                               │
+                                                                             │           11       5555555                                                                                                                                         │                                                                                                                                              4                                │
+                                                                             │          111    555                                                                                                                                                │                                                                                                                                             4                                 │                                                                                    ...
+                                                                             │          1 1    5                                                                                                                                                  │                                                                                                                                            4                                  │                                                                                 .... ...
+                                                                             │         1  1     5                                                                                                                                                 │                                                                                                                                           44        44                        │                                                                               ...       .
+                                                                             │         1 11     5                                                                                                                                                 │                                                                                                                                           4   44444444                        │                                                                               ..        .
+                                                                             │        1  1      5                                                                                                                                                 │                                                                                                                                           44444     44                        │                                                                                 .       .
+                                                                             │       1   1     5                                                                                                                                                  │                                                                                                                                                     4                         │                                                                                 .      .
+                                                                             │      1    1     55555555                                                                                                                                           │                                                                                                                                                     4                         │                                                                                   .  ..
+                                                       .... .                │     1     1     5      555                                                                                                                                         │                                                                                                                                                     4                         │                                                                                    ..
+                                                       .... ..               │    11    1               55                                                                                                                                        │                                                                                                                                                    4                          │                          . .  . .
+                                                      . .    .               │          1                5                                                                                                                                        │                                                                                                                                                    4                          │                       .          ..
+       ...         ..                                   .    .               │          1                5                                                                                                                                        │                                                                                                                                                    4                          │                     .             ..
+     ..    ..        . .                                .    .               │          1                5                                                                                                                                        │                                                                                                                                                    4                          │                     .            ...                     .. .
+    .        .           . .     ..                      .....               │          1                5                                                                                                                                        │                                                                                                                                                    4                          │                     .             ..              .... .     .
+   .          .               . ..             .. ...                        │          1     55        5                                                                                                                                         │                                                                                                                                                    4                          │                     .             ..            .  .           .
+   .          .                  .         . ..      .                       │          1      5555  5555                                                                                                                                         │                                                                                                                                                   4                           │                                  ..            . .              .
+    .        .                            ..         .                       │          1          555                                                                                                                                            │                                                                                                                                                   44                          │                      .         ..                 ..             .
+             .                  .        ...          .                      │          1                                                                                                                                                         │                                                                                                                                                    4                          │                       . .. . .   .            .                  .
+            .                   .       .              .                     │                                                                                                                                                                    │                                                                                                                                                                               │                                  .
+          .            ... .  . .  ...  .              .                     │                                                                                                                                                                    │                                                                                                                                                                               │                                 .            .                   .
+         .                                                                   │                                                                                                                                                                    │                                                                                                                                                                               │                                              .                   .
+       .                        .       .              .                     │                                                                                                                                                                    │                                                                                                                                                                               │                                 .                                .
+ ..   .                        .        .             .                      ├────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┤                      .         .             .                 .
+......          .              .         .           .                       │                                                                                                                                                                    │                                                                                                                                                                               │                      .
+..     .       ..             .           ..      . .                        │                                                                                                                                                                    │                                                                                                                                                                               │                      .        ..              .              ..
+          . .  .              .            . .. .                            │                                                                                                                                                                    │                                                                                                                                                                               │                      .      ..                 .          .
+             ..                                                              │                                                                                                                                                                    │                                                                                                                                                                               │                       ......                    ..     ..
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │                                                   .. .
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                    55555555555555             │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                    5                          │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                   5                           │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                   5                           │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                  5                            │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 5                             │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 5                             │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                55555                          │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                5    5555                      │
+                                                                             │           1                                                                                                                                                        │                                                                                                                                                         55                    │
+                                                                             │          11       4                                                                                                                                                │                                                                                                                                                          55                   │
+                                                                             │          11       4                                                                                                                                                │                                                                                                                                                           5                   │
+                                                                             │         111      44                                                                                                                                                │                                                                                                                                                           5                   │
+                                                                             │         1 1      4                                                                                                                                                 │                                                                                                                                                           5                   │
+                                                                             │       11  1      4                                                                                                                                                 │                                                                                                                                                           5                   │
+                                                                             │       1   1     4                                                                                                                                                  │                                                                                                                                                           5                   │
+                                                                             │     11    1     4                                                                                                                                                  │                                                                                                                                                          5                    │
+                                                                             │     1     1    4                                                                                                                                                   │                                                                                                                                                          5                    │
+                                                                             │           1    44444444                                                                                                                                            │                                                                                                                                             555         5                     │
+                                                                             │           1          44                                                                                                                                            │                                                                                                                                                5555   55                      │
+                                                                             │           1          4                                                                                                                                             │                                                                                                                                                    555                        │
+                                                                             │                      4                                                                                                                                             │                                                                                                                                                                               │
+                                                                             │           1          4                                                                                                                                             │                                                                                                                                                                               │
+                                                                             │           1          4                                                                                                                                             │                                                                                                                                                                               │
+                                                                             │          1           4                                                                                                                                             │                                                                                                                                                                               │
+                                                                             │          1           4                                                                                                                                             │                                                                                                                                                                               │
+                                                                             │          1          4                                                                                                                                              │                                                                                                                                                    6666                       │
+                                                                             │                     4                                                                                                                                              │                                                                                                                                                  666                          │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                 66                            │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                6                              │
+                                                                             │                                                                                                                                                                    │                                                                                                                                              66                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                             66                                │
+                                                                             │                                                                                                                                                                    │                                                                                                                                            66                                 │
+                                                                             │                                                                                                                                                                    │                                                                                                                                            6                                  │
+                                                                             │                                                                                                                                                                    │                                                                                                                                           6                                   │
+                                                                             │         1                                                                                                                                                          │                                                                                                                                           6                                   │
+                                                                             │        11     333333                                                                                                                                               │                                                                                                                                          6                                    │
+                                                                             │        11          33                                                                                                                                              │                                                                                                                                          6   66666666666                      │
+                                                                             │       1 1          33                                                                                                                                              │                                                                                                                                          6 666         6666                   │
+                                                                             │      11 1         33                                                                                                                                               │                                                                                                                                          666              6                   │
+                                                                             │      1  1        33                                                                                                                                                │                                                                                                                                          66               66                  │
+                                                                             │     1   1      333                                                                                                                                                 │                                                                                                                                          6                 6                  │
+                                                                             │    11   1        3333                                                                                                                                              │                                                                                                                                          6                 6                  │
+                                                                             │    1    1           3                                                                                                                                              │                                                                                                                                          6                 6                  │
+                                                                             │         1           3                                                                                                                                              │                                                                                                                                           6               66                  │
+                                                                             │         1           3                                                                                                                                              │                                                                                                                                           66             66                   │
+                                                                             │         1          33                                                                                                                                              │                                                                                                                                            666         666                    │
+                                                                             │         1          3                                                                                                                                               │                                                                                                                                               6666666666                      │
+                                                                             │         1         3                                                                                                                                                │                                                                                                                                                                               │
+                                                                             │         1     3333                                                                                                                                                 │                                                                                                                                                                               │
+                                                                             │         1     3                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                        9999                                                                                                                                   │
+                                                                             │                                                           1                                                                                                        │                                     999    999                                              888888                                                                            │
+                                                                             │                                                           1                                                      1                                                 │                                    99       99                                            88     8888                                   7                                     │
+                                                                             │                                                          11            1                                        11                                                 │                                   99        99                                           88         88                                   7777777777     7                     │
+                                                                             │         11                                               11           11                                       111        0000                                     │                                   9         9                                            8           88                                            777777                     │
+                                                                             │         1    2222                                      111           1 1                                      11 1       000  0000                                 │                                   99       99                                            88           8                                                 7                     │
+                                                                             │        11    2   2                                    11 1          1 1                                      11  1      00       000                               │                                    9999999999                                              88        88                                                 7                     │
+                                                                             │       111        2                                    1  1         1  1                                     11   1      0          00                              │                                            9                                                8888     8                                                 77                     │
+                                                                             │      11 1       22                                   1   1       11   1                                    11    1     00            0                             │                                            9                                                   88   8                                                  7                      │
+                                                                             │    111  1      2                                    1    1      1     1                                  111    1      0             0                             │                                            9                                                     888                                                   7                      │
+                                                                             │        1      22                                   1    1      11     1                                         1      0             0                             │                                            9                                                    8888                                                   7                      │
+                                                                             │        1     22                                         1             1                                         1      0             0                             │                                           99                                                   8    8                                                  7                      │
+                                                                             │        1    22                                          1             1                                         1      0            00                             │                                           9                                                  88     88                                        77777777777777777               │
+                                                                             │        1   22                                           1            1                                          1       0           0                              │                                9          9                                                88        8                                                77                      │
+                                                                             │       11  22                                           1             1                                          1       0          00                              │                                 99       9                                                88         8                                                7                       │
+                                                                             │       1  2222                                          1             1                                          1        0        00                               │                                   999 999                                                8           8                                                7                       │
+                                                                             │       1     2222   2                                                 1                                          1         000   000                                │                                      9                                                   88         88                                                7                       │
+                                                                             │       1        22222                                                                                            1            000                                   │                                                                                            88      88                                                 7                       │
+                                                                             │                                                                                                                 1                                                  │                                                                                              8888888                                                  7                       │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                       7                       │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                       7                       │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             │                                                                                                                                                                    │                                                                                                                                                                               │
+                                                                             └────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┘
+                                                                                                                                                                                                                                                  │                                                                                                                                                                                                                                                     . ..
+                                                                                                                                                                                                                                                                                                                                                                                                                                                .                                                      .. ..
+                                                                                                                                                                                                                                                                                                                                                                                                                                                .          .  ...                                     ..    ....
+                                                                                                                                                                                                                                                                                                                                                                                                                                                .     . ..       .        ....                        ..       .
+                                           . ..                         ...                                                                                                                                                                                                                                                                                                                                                                    .     .           .        .    . ......               .       .
+                                .          .   ..                      ......                                                                                                                                                                                                                                                                                                                                                                  .                .         .                           .      .
+                            ...... .            .                      .     .                                                                                                                                                      . ...                                                                                                                                                                                                      .               .          .                           .      .
+                                   .            .              ..      .     .                                                                                                                                              ..     .     . .                         ... .                                                                                                                                                                                    ..          .                            ..  ..
+                                    .           .       ........       .     .                                                                                                                                              .     .        .            . .         ..    .                                                                                                                                                                   .             ..            .                               ..
+                                   .            .         ..           .    ..                                                                                                                                              .     ..     . .         . .   ..       ..    ..                                                                                                                                                                 .             .... ...       .
+                                  .            .            .           .....                                                                                                                                              ..       .  ...         .....    .        .    .                                                                                                                                                                 . .                   ..     ... ... .
+                            ..   .             .             .                                                                                                                                                                       ..            ..        .        .. .                                                                                                                                                                 .                       .     ..       .
+                          ... . ..            ....            .                                                                                                                                                           ..        . .           . .        .                                                                                                                                                                             .  .       .           .               .
+                          ... ...           ...   .      .    .                                                                                                                                                          . .       .   .          .  .       .                                                                                                                                                                                .       . .         .       .       .
+                                 .. ....           .......    .                                                                                                                                                         ..        .     .         .          .                                                                                                                                                                               .           ..     ..        ..      .
+                                                         .   .                                                                                                                                                         .   .      .      .        .          .                                                                                                                                                                               .              ....           ..    ..
+                                                         . ...                                                                                                                                                         .  .        .     .        .          .                                                                                                                                                                              ..                                ...
+                                                                                                                                                                                                                          .         .     .                 .                                                                                                                                                                               .
+                                                                                                                                                                                                                          . ..        ......       .       .
+                                                                                                                                                                                                                          .                        . .    .
+                                                                                                                                                                                                                                                      ...
+*/
+
 interface IProps extends IUIComponent {
   children?: any
   orientation?: Range<360>
@@ -12,143 +233,58 @@ interface IProps extends IUIComponent {
   [key: string]: any
 }
 
-const Sun = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="74"
-      height="74"
-      viewBox="0 0 74 74"
-    >
-      <g
-        id="Group_223"
-        data-name="Group 223"
-        transform="translate(-494.5 -732.5)"
-      >
-        <g id="Group_222" data-name="Group 222" transform="translate(497 735)">
-          <circle
-            id="Ellipse_11"
-            data-name="Ellipse 11"
-            cx="15.5"
-            cy="15.5"
-            r="15.5"
-            transform="translate(20 20)"
-            fill="#fff5cb"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_57"
-            data-name="Line 57"
-            y2="6"
-            transform="translate(35 1)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_58"
-            data-name="Line 58"
-            y2="5"
-            transform="translate(35 63)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_59"
-            data-name="Line 59"
-            x2="4.376"
-            y2="4.376"
-            transform="translate(10.924 10.924)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_60"
-            data-name="Line 60"
-            x2="4.376"
-            y2="4.376"
-            transform="translate(54.502 54.502)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_61"
-            data-name="Line 61"
-            x2="6"
-            transform="translate(1 35)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_62"
-            data-name="Line 62"
-            x2="5"
-            transform="translate(63 35)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_63"
-            data-name="Line 63"
-            y1="4.376"
-            x2="4.376"
-            transform="translate(10.924 54.502)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-          <line
-            id="Line_64"
-            data-name="Line 64"
-            y1="4.376"
-            x2="4.376"
-            transform="translate(54.502 10.924)"
-            fill="none"
-            stroke="#fff5cb"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="7"
-          />
-        </g>
-        <text
-          id="S"
-          transform="translate(525 778)"
-          fill="#edc749"
-          stroke="rgba(0,0,0,0)"
-          stroke-width="1"
-          font-size="25"
-          font-family="Karmina-Bold, Karmina"
-          font-weight="700"
-        >
-          <tspan x="0" y="0">
-            S
-          </tspan>
-        </text>
-      </g>
-    </svg>
-  )
+/*
+    This function will return the position of the Sun, based on the degrees given
+    Check the drawing above for reference.
+    
+    Full range:
+    (0-20) => 0
+    (20-40) => 1
+    (40-50) => 2 (diagonal)
+    (50-70) => 3
+    (70-90) => 4
+    (90-110) => 5
+    (110-130) => 6
+    (130-140) => 7 (diagonal)
+    (140-160) => 8
+    (160-180) => 9
+    (180-200) => 10
+    (200-220) => 11
+    (220-230) => 12 (diagonal)
+    (230-250) => 13
+    (250-270) => 14
+    (270-290) => 15
+    (290-310) => 16
+    (310-320) => 17 (diagonal)
+    (320-340) => 18
+    (340-359) => 19
+    
+*/
+const findPos = (orientation: Range<360>): number => {
+  return Math.floor((19 / 360) * orientation)
+}
+
+const getLocation = (
+  pos: number
+): {
+  top?: string | undefined
+  bottom?: string | undefined
+  left?: string | undefined
+  right?: string | undefined
+} => {
+  if (pos == 0) {
+    return {
+      top: '0',
+      left: '55%'
+    }
+  }
+  // and so on..
+  // needs to be completed
+  else if (pos == 1) {
+  }
+  return {
+    top: '0'
+  }
 }
 
 export function Sunshine({
@@ -157,8 +293,155 @@ export function Sunshine({
   tooltip = 'Sun'
 }: IProps) {
   return (
-    <Box>
-      <Sun />
+    <Box
+      style={{
+        width: '100%',
+        height: '100%',
+        background: `linear-gradient(${orientation}deg, rgba(2,0,36,0) 0%, rgba(237,206,126,0.7) 100%)`
+      }}
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="74"
+        height="74"
+        viewBox="0 0 74 74"
+        style={{
+          position: 'relative',
+          ...getLocation(findPos(orientation))
+        }}
+      >
+        <g
+          id="Group_223"
+          data-name="Group 223"
+          transform="translate(-494.5 -732.5)"
+        >
+          <g
+            id="Group_222"
+            data-name="Group 222"
+            transform="translate(497 735)"
+          >
+            <circle
+              id="Ellipse_11"
+              data-name="Ellipse 11"
+              cx="15.5"
+              cy="15.5"
+              r="15.5"
+              transform="translate(20 20)"
+              fill="#fff5cb"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_57"
+              data-name="Line 57"
+              y2="6"
+              transform="translate(35 1)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_58"
+              data-name="Line 58"
+              y2="5"
+              transform="translate(35 63)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_59"
+              data-name="Line 59"
+              x2="4.376"
+              y2="4.376"
+              transform="translate(10.924 10.924)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_60"
+              data-name="Line 60"
+              x2="4.376"
+              y2="4.376"
+              transform="translate(54.502 54.502)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_61"
+              data-name="Line 61"
+              x2="6"
+              transform="translate(1 35)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_62"
+              data-name="Line 62"
+              x2="5"
+              transform="translate(63 35)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_63"
+              data-name="Line 63"
+              y1="4.376"
+              x2="4.376"
+              transform="translate(10.924 54.502)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+            <line
+              id="Line_64"
+              data-name="Line 64"
+              y1="4.376"
+              x2="4.376"
+              transform="translate(54.502 10.924)"
+              fill="none"
+              stroke="#fff5cb"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="7"
+            />
+          </g>
+          <text
+            id="S"
+            transform="translate(525 778)"
+            fill="#edc749"
+            stroke="rgba(0,0,0,0)"
+            strokeWidth="1"
+            fontSize="25"
+            fontFamily="Karmina-Bold, Karmina"
+            fontWeight="700"
+          >
+            <tspan x="0" y="0">
+              S
+            </tspan>
+          </text>
+        </g>
+      </svg>
     </Box>
   )
 }
